Creacion de vistas y sistemas de para funcionamiento privado y publico en las tablas adopciones y mascotas

+++ database/migrations/2025_11_07_201111_create_administradores_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('Telefono')->unique();
             $table->string('Email')->unique();
             $table->string('Password_admin');
-            $table->Float('Sueldo_admin');
+            $table->decimal('Sueldo_admin', 10, 2);
             $table->timestamps();
         });
     }
